# openapi.json
{"openapi": "3.0.2", "info": {"title": "Document Loader Connector", "version": "0.1.0"}, "paths": {"/api/v1/load/": {"post": {"tags": ["document loader"], "summary": "Load Document", "description": "This endpoint is used to load any type of document from a local filesystem location", "operationId": "load_document_api_v1_load__post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_load_document_api_v1_load__post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Load Document Api V1 Load  Post"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/load/remote": {"post": {"tags": ["document loader"], "summary": "Load Remote Document", "description": "This endpoint is used to load any type of document from a remote location", "operationId": "load_remote_document_api_v1_load_remote_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/BaseLoader"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Load Remote Document Api V1 Load Remote Post"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"BaseLoader": {"title": "BaseLoader", "required": ["url"], "type": "object", "properties": {"url": {"title": "Url", "type": "string"}}}, "Body_load_document_api_v1_load__post": {"title": "Body_load_document_api_v1_load__post", "required": ["file"], "type": "object", "properties": {"file": {"title": "File", "type": "string", "format": "binary"}}}, "HTTPValidationError": {"title": "HTTPValidationError", "type": "object", "properties": {"detail": {"title": "Detail", "type": "array", "items": {"$ref": "#/components/schemas/ValidationError"}}}}, "ValidationError": {"title": "ValidationError", "required": ["loc", "msg", "type"], "type": "object", "properties": {"loc": {"title": "Location", "type": "array", "items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}}, "msg": {"title": "Message", "type": "string"}, "type": {"title": "Error Type", "type": "string"}}}}}}